requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
ffmpeg-python>=0.2.0
deep-sort-realtime>=1.3.2
pydantic>=2.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0
langgraph>=0.1.0
click>=8.1.0
imageio-ffmpeg>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
roboflow>=1.1.0
easyocr>=1.7.0
anthropic>=0.40.0
openai>=1.50.0
